Refs #2974 Update and fix several errors in the unit tests

+++ test/blackbox/RTPSWithRegistrationReader.hpp
@@ -280,7 +280,7 @@ class RTPSWithRegistrationReader
         {
             eprosima::fastrtps::rtps::Locator_t loc;
             loc.set_IP4_address(ip);
-            loc.set_port(port);
+            loc.set_port(static_cast<uint16_t>(port));
             reader_attr_.endpoint.multicastLocatorList.push_back(loc);
 
             return *this;
